requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
google-generativeai==0.3.1
python-dotenv==1.0.1
pytest==8.0.2
gunicorn==21.2.0
requests==2.31.0 